list working

+++ app/src/main/java/com/sgut/android/beerapplication/data/remote/responses/Amount.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Amount (
 
-  @SerializedName("value" ) var value : Int?    = null,
+  @SerializedName("value" ) var value : Double?    = null,
   @SerializedName("unit"  ) var unit  : String? = null
 
 )
